requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
Pillow==10.0.1
itsdangerous==2.1.2
python-dotenv==1.0.1 